:sparkles: feat: Distance 계산 기능 추가, Double형으로 Number 멤버 변경

+++ src/main/java/org/cleancode/calculator/calcuables/Number.java
@@ -8,14 +8,14 @@ import java.util.Objects;
 
 @HandleFor(target = CalculablePattern.NUMBER)
 public class Number implements Calculable {
-    private final int value;
+    private final double value;
     public static Number of(String item){
-        return new Number(Integer.parseInt(item));
+        return new Number(Double.parseDouble(item));
     }
-    public Number(int value){
+    public Number(double value){
         this.value = value;
     }
-    public int getValue() {
+    public double getValue() {
         return value;
     }
     @Override
